BT-adapter/requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
annotated-types==0.7.0
argcomplete==3.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==12.1.0
beaker-gantry==1.8.4
beaker-py==1.32.0
boto==2.49.0
cachetools==5.5.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-help-colors==0.9.4
cmake==3.29.5
colorama==0.4.6
colorful==0.5.6
contourpy==1.2.1
coverage==7.5.3
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
decorator==4.4.2
decord==0.6.0
docker==7.1.0
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.19
filelock==3.14.0
flake8==7.0.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
ftfy==6.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gcs-oauth2-boto-plugin==3.2
gitdb==4.0.11
GitPython==3.1.43
google-apitools==0.5.32
google-auth==2.17.0
google-auth-httplib2==0.2.0
google-reauth==0.1.1
greenlet==3.0.3
grpcio==1.65.2
gsutil==5.31
httplib2==0.20.4
huggingface-hub==0.17.3
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
imgaug==0.4.0
importlib_metadata==7.1.0
iniconfig==2.0.0
interrogate==1.7.0
iopath==0.1.10
ipython==8.25.0
isort==4.3.21
jedi==0.19.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
line_profiler==4.1.3
lit==18.1.7
llvmlite==0.42.0
lmdb==1.4.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mmaction2==1.0.0
mmcv==2.0.1
mmengine==0.10.4
model-index==0.1.11
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.1.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
opencv-python-headless==4.10.0.82
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.0
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
petname==2.6
pexpect==4.9.0
pillow==10.3.0
PIMS==0.7
pip==24.0
platformdirs==4.2.2
pluggy==1.5.0
pooch==1.8.2
portalocker==2.8.2
proglog==0.1.10
prompt_toolkit==3.0.47
propcache==0.2.0
protobuf==4.25.4
psutil==5.9.8
ptyprocess==0.7.0
pudb==2024.1
pure-eval==0.2.2
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
pynvim==0.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pytest==8.2.2
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
pytorchvideo==0.1.5
PyTurboJPEG==1.7.3
pytz==2023.4
pyu2f==0.1.5
PyYAML==6.0.1
regex==2024.5.15
requests==2.28.2
retry-decorator==1.1.1
rich==13.4.2
rsa==4.7.2
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
sentry-sdk==2.5.1
setproctitle==1.3.3
setuptools==60.2.0
shapely==2.0.4
six==1.16.0
slicerator==1.1.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
sympy==1.12.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.22
timm==1.0.3
tokenizers==0.14.1
tomli==2.0.1
torch==2.0.0
torch-tb-profiler==0.4.3
torchvision==0.15.1
tqdm==4.65.2
traitlets==5.14.3
transformers==4.35.0
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.18
urwid==2.6.14
urwid_readline==0.14
wandb==0.17.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
xdoctest==1.1.5
yacs==0.1.8
yapf==0.40.1
yarl==1.16.0
zipp==3.19.2
